+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/application/LoadTaskSpec.groovy
@@ -5,8 +5,12 @@ package com.yahoo.bard.webservice.application
 import static spock.util.matcher.HamcrestMatchers.closeTo
 import static spock.util.matcher.HamcrestSupport.expect
 
+import com.yahoo.bard.webservice.data.dimension.TimeoutException
 import com.yahoo.bard.webservice.druid.client.FailureCallback
 import com.yahoo.bard.webservice.druid.client.HttpErrorCallback
+import com.yahoo.bard.webservice.logging.TestLogAppender
+
+import com.fasterxml.jackson.databind.ObjectMapper
 
 import spock.lang.Ignore
 import spock.lang.Specification
@@ -17,6 +21,8 @@ import java.util.concurrent.TimeUnit
 
 @Timeout(30) // Fail test if hangs
 class LoadTaskSpec extends Specification {
+
+    TestLogAppender logAppender = new TestLogAppender()
 
     // All times are in millis unless specified otherwise
     int expectedDelay = 500
@@ -42,7 +48,7 @@ class LoadTaskSpec extends Specification {
         }
 
         @Override
-        void run() {
+        void runInner() {
             start = System.currentTimeMillis()
             sleep(expectedDuration)
             end = System.currentTimeMillis()
@@ -60,7 +66,7 @@ class LoadTaskSpec extends Specification {
         }
 
         @Override
-        void run() {
+        void runInner() {
             start = System.currentTimeMillis()
             sleep(expectedDuration)
             end = System.currentTimeMillis()
@@ -230,5 +236,43 @@ class LoadTaskSpec extends Specification {
 
         cleanup:
         scheduler.shutdownNow()
+    }
+
+    class FailingTestLoadTask extends LoadTaskSpec.OneOffTestLoadTask {
+
+
+        public static final String UNIQUE_VALUE_1 = "UNIQUE_VALUE_1"
+
+        FailingTestLoadTask(long delay) {
+            super(delay)
+        }
+
+        @Override
+        void runInner() {
+            throw new TimeoutException(UNIQUE_VALUE_1);
+        }
+    }
+
+
+    def "Test error gets logged on failed test"() {
+        setup:
+        setup: "Instantiate a task scheduler"
+        ObjectMapper MAPPER = new ObjectMappersSuite().getMapper()
+
+
+        TaskScheduler scheduler = new TaskScheduler(2)
+        LoadTask<Boolean> loader = new FailingTestLoadTask(expectedDelay)
+        loader.setFuture(
+                scheduler.schedule(
+                        loader,
+                        1,
+                        TimeUnit.MILLISECONDS
+                )
+        )
+        Thread.sleep(10)
+
+        expect:
+        loader.failed
+        logAppender.getMessages().find { it.contains(/UNIQUE_VALUE_1/) }
     }
 }